pyproject.toml
[project]
name = "trusted-services-loadtests"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aio-pika",
    "asyncpg",
    "configargparse",
    "flask",
    "flask-cors",
    "flask-login",
    "locust",
    "locust-cloud",
    "python-dotenv",
    "python-socketio",
    "python-engineio",
    "redis",
    "requests",
    "psycopg[binary,pool]",
    "pika",
]

[tool.locust]
# locustfile = "locust_files/my_locust_file.py"
headless = true
# master = true
# expect-workers = 5
host = "http://localhost"
users = 100
spawn-rate = 10
run-time = "10m"
# tags = ["Critical", "Normal"]
csv = "results"
